[Issue 5585][pulsar-client] Fix producer Semaphore release error  (#5587)
### Motivation

Fixes #5585 

This PR is try to fix the dead lock issue which mentioned in #5585. The core problem is the semaphore release count calculation in replicate scenario.

Replicator send a batch message(from the original producer) without batcher, so that the batch message acquire 1 each batch message but release it by number of messages in the batch. This PR will handle the semaphore release in replicate scenario.

### Modifications

Release only one semaphore in replicate scenario.