Specify namespace when launching pulsar-manager (#7037)
In the `Get started in Kubernetes` quickstart tutorial, one of the commands is insufficient and throws an error. I have proposed a solution to this by specifying the `pulsar` namespace to the command to fix the issue. There is no associated issue that I created as the work and fix is rather trivial.