fix NullPointerException with SubscriptionMode.NonDurable (#7118)
Fixes #7046


### Motivation

This code is to enable non-persistent subscriptions to continue to consume， based on the Id of the first message in the client's ReceiverQueue when reconnection is triggered.

But it's not considered the case that there is no messageId for the first connection. Therefore, we need to determine whether the MessageId is null.