Close producer without state check while the topic does not exists. (#7120)
Fixes #6838 related to #6879

### Motivation

Close producer without state check while the topic does not exists. The problem is the `State` of the producer failed to pass the check condition so that the producer can't be closed. If TopicDoesNotExsitsException happens on a producer, we can directly close this producer without state checking for the producer.