passes on authenticationData (#7694)
# Fixes #7693

Motivation
When trying to chase down a bug in my custom AuthenticationProvider I found that when attempting to consume a topic that I wasn't authorized on would fail to produce an unauthorized response while using the pulsar-client. I noticed that the canConsumeAsync within AuthorizationService canLookUpAsync doesn't pass along the AuthenticationDataSource when it should. My code relied on that to be passed down each time and thus my methods would never return.

# Modifications
I've added the AuthenticationDataSource to the canConsumeAsync method.